+++ .scss-lint.yml
@@ -4,9 +4,9 @@ linters:
   LeadingZero:
     style: exclude_zero
   NestingDepth:
-    max_depth: 3
+    max_depth: 5
   SelectorDepth:
-    max_depth: 3
+    max_depth: 5
   SelectorFormat:
     convention: hyphenated_BEM
   StringQuotes:
