+++ app/assets/stylesheets/common/components/card.scss
@@ -1,5 +1,5 @@
 .c-card {
-  background-color: $white;
+  background-color: $color-white;
   border-radius: 2px;
-  box-shadow: 0 1px 2px $cadet-gray;
+  box-shadow: 0 1px 2px $color-cadet;
 }
