+++ app/assets/stylesheets/common/components/watchers-count.scss
@@ -1,0 +1,12 @@
+.c-watchers-count {
+  line-height: 1;
+  margin-bottom: .25rem;
+
+  &__number {
+    margin-right: .25rem;
+  }
+
+  &__text {
+    font-size: .9rem;
+  }
+}
