+++ app/assets/stylesheets/common/foundations/bootstrap/utilities/background.scss
@@ -1,0 +1,4 @@
+.bg-transparent {
+  background-color: transparent;
+  color: $color-white;
+}
