+++ app/assets/stylesheets/common/foundations/variables.scss
@@ -1,10 +1,13 @@
-$mine-shaft-black: #404040;
-$river-black: #5f697a;
+$color-mine-shaft: #404040;
+$color-river: #5f697a;
 
-$cadet-gray: rgba(43, 59, 93, .29);
+$color-cadet: rgba(43, 59, 93, .29);
 
-$haze-white: #f9f9fb;
-$white: #fff;
+$color-haze: #f9f9fb;
+$color-white: #fff;
+
+$color-java: #3fc1c9;
+$color-san-juan: #466c95;
 
 $color-mizuho: #f85b73;
 $color-kei: #c6baa2;
