+++ app/assets/stylesheets/db/layouts/sidebar.scss
@@ -1,5 +1,5 @@
 .l-sidebar {
-  background-color: $mine-shaft-black;
+  background-color: $color-mine-shaft;
   float: left;
   height: 100%;
   overflow-y: auto;
@@ -8,10 +8,10 @@
   width: 14rem;
 
   a {
-    color: $white;
+    color: $color-white;
 
     &:hover {
-      color: $white;
+      color: $color-white;
       text-decoration: none;
     }
