+++ app/assets/stylesheets/db/layouts/topbar.scss
@@ -1,6 +1,6 @@
 .l-topbar {
-  background-color: $white;
-  box-shadow: 0 2px 2px -2px $cadet-gray;
+  background-color: $color-white;
+  box-shadow: 0 2px 2px -2px $color-cadet;
   height: 4.5rem;
   line-height: 4.5rem;
