+++ app/assets/stylesheets/pc.scss
@@ -1,3 +1,7 @@
 //= require_self
 
 @import "common/foundations/foundation";
+@import "common/components/**/*";
+
+@import "pc/components/**/*";
+@import "pc/projects/**/*";
