+++ app/assets/stylesheets/pc/components/search-form.scss
@@ -1,0 +1,17 @@
+.c-search-form {
+  padding: .08rem 0;
+
+  .form-control {
+    background-color: rgba(0, 0, 0, .1);
+    border: 1px solid rgba(255, 255, 255, .15);
+    border-color: rgba(255, 255, 255, .15);
+    color: $color-white;
+    width: 280px;
+
+    &:focus {
+      background-color: rgba(0, 0, 0, .1);
+      border-color: rgba(255, 255, 255, .15);
+      color: $color-white;
+    }
+  }
+}
