+++ app/assets/stylesheets/pc/projects/top-hero.scss
@@ -1,0 +1,28 @@
+.p-top-hero {
+  height: 400px;
+  position: relative;
+
+  &__bg-image {
+    @include filter(brightness(50%));
+    background-position: center;
+    background-size: cover;
+    height: 100%;
+  }
+
+  &__content {
+    height: 100%;
+    left: 0;
+    position: absolute;
+    top: 0;
+    width: 100%;
+
+    h1 {
+      color: $color-white;
+      margin-top: 8rem;
+    }
+  }
+
+  &__description {
+    color: $color-white;
+  }
+}
