+++ app/assets/stylesheets/pc/projects/top-work.scss
@@ -1,0 +1,13 @@
+.p-top-work {
+  @include display(flex);
+
+  &__left {
+    @include flex-basis(250px);
+    @include flex-shrink(0);
+    line-height: 0;
+
+    a {
+      margin-bottom: .5rem;
+    }
+  }
+}
