+++ app/controllers/home_controller.rb
@@ -9,7 +9,6 @@ class HomeController < ApplicationController
     @top_work = GuestTopPageService.top_work
     @works = GuestTopPageService.works
     @cover_image_work = GuestTopPageService.cover_image_work
-    @activities = GuestTopPageService.activities unless browser.mobile?
 
     render :index_guest
   end
