+++ app/models/season.rb
@@ -20,8 +20,7 @@
 class Season < ActiveRecord::Base
   has_many :works
 
-  delegate :humanize_name, to: :decorate
-  delegate :yearly_season_ja, to: :decorate
+  delegate :local_name, to: :decorate
 
   NAME_DATA = {
     winter: "冬",
