--- app/views/application/_error_panel.html.slim
@@ -1,7 +1,0 @@
-- if resource.errors.any?
-  .alert.alert-danger
-    h4.alert-heading
-      = t "words.error"
-    ul
-      - resource.errors.full_messages.each do |msg|
-        li = msg
