+++ app/views/application/components/_footer.html.slim
@@ -1,0 +1,32 @@
+.c-footer.container
+  .row
+    .col-xs-3
+    .col-xs-3
+      h4.font-weight-bold = t "words.contents"
+      ul.list-unstyled
+        li = link_to t("head.title.pages.about"), about_path
+        li = link_to t("words.annict_api"), "https://annict.wikihub.io/wiki/api", target: "_blank"
+        li = link_to t("words.terms_of_use"), terms_path
+        li = link_to t("words.privacy_policy"), privacy_path
+    .col-xs-3
+      h4.font-weight-bold = t "words.community"
+      ul.list-unstyled
+        li = link_to t("words.blog"), "http://blog.annict.com", target: "_blank"
+        li = link_to "Twitter", "https://twitter.com/anannict", target: "_blank"
+        li = link_to "GitHub", "https://github.com/annict", target: "_blank"
+        li = link_to "Slack", "http://slack.annict.com", target: "_blank"
+    .col-xs-3
+      h4.font-weight-bold = t "words.seasonal_anime"
+      - next_season_slug = ENV.fetch("ANNICT_NEXT_SEASON")
+      - year, name = next_season_slug.split("-")
+      - years = year.to_i.downto(2000).to_a
+      - name_data = Season::NAME_DATA
+      - names = name_data.keys.map(&:to_s).reverse
+      - slugs = years.map { |y| names.map { |sn| "#{y}-#{sn}" } }.flatten
+      - index = slugs.index(next_season_slug)
+      - slugs = slugs[index..index + 4]
+      ul.list-unstyled
+        - slugs.each do |slug|
+          li
+            - year, name = slug.split("-")
+            = link_to season_local_name(year, name), season_works_path(slug: slug)
