+++ app/views/application/components/_sign_up_modal.html.slim
@@ -1,0 +1,5 @@
+.c-sign-up-modal.modal
+  .modal-dialog.modal-lg
+    .modal-content
+      .modal-body
+        | (^o^)
