+++ app/views/db/works/index.html.slim
@@ -20,7 +20,7 @@
         = t "words.season"
       .dropdown-menu.pre-scrollable
         - Season.order(sort_number: :desc).each do |season|
-          = link_to season.decorate.humanize_name, season_db_works_path(slug: season.slug), class: "dropdown-item"
+          = link_to season.decorate.local_name, season_db_works_path(slug: season.slug), class: "dropdown-item"
     li.nav-item
       = active_link_to t("resources.episode.no_episodes"), resourceless_db_works_path(name: :episode), class: "nav-link", active: (params[:name] == "episode")
     li.nav-item
